Create README.md
İBBCOİN
IBBCOIN IS FOR WEB3 ECONOMY.
Culture has found new expression on the web3 through the arts, games, entertainment and events. The possibilities of the impact of the blockchain on culture are so endless that all of them cannot yet be predicted. IMM is a community controlled and created token made to support what's next. It will serve as a decentralized protocol layer for community-led initiatives that drive culture to metadata.

IMM FOUNDATION
İBB Foundation is the steward of İBBCoin. This is not a supervisor, it is the base layer that IBBCoin holders in the IBBCoin DAO can build on.

The Foundation facilitates decentralized and community-led administration and is designed to become more decentralized over time. IbbCoin is tasked with managing the decisions of the DAO and is responsible for day-to-day management, bookkeeping, project management and other duties ensuring the DAO community has the support they need to make their ideas come true.

The aim of the IMM Foundation is to manage the growth and development of the IMM ecosystem in a fair and inclusive way. It uses the multisig wallet-controlled Ecosystem Fund to pay for its expenses as directed by the ibbCoin DAO and provides an infrastructure for ibbCoin holders to collaborate through open and permissionless governance processes.

KNIFE
The IbbCoin DAO exists because decentralized governance is critical to building and managing a globally dispersed community and therefore critical to the success of the IBB ecosystem.

The IMM Improvement Proposal Process (see Management ) will allow IBBCoin DAO members to make decisions regarding Ecosystem Fund allocations, governance rules, projects, partnerships and more. İBBCoin DAO membership is open to all İBBCoin owners.

Governance is critical for building and managing a globally dispersed community and therefore for the success of the IMM ecosystem. 

PANO
A special council on the IBB Foundation (DAO's “Board”) provides oversight of Foundation executives at the behest of IBBCoin DAO members. The board's purpose is to manage DAO offerings and serve the community's vision. It is collected in proposals that require administrative review within the scope of İBBCoin DAO rules. The first Board serves for a six-month term, after which DAO members vote annually for Board members.

IBBCOIN PROTOCOL
IbbCoin is a BEP20 governance and utility token used in the IBB ecosystem to empower a decentralized community structure at the forefront of web3.

As the open-source protocol layer of the ecosystem, IBBCoin serves several purposes:



IBBCOIN ALLOCATION
The total supply of ibbCoin is permanently fixed at 100,000,000,000,000 coins. No ability to print money arises through the contract interface, so the total supply never increases. Similarly, the contract interface does not show any token writing capacity, so the total supply will never decrease.

DOES THE FOUNDATION CONTROL IBBCOIN AND/OR IBBCOIN DAO? IF NOT, WHAT ROLE DO THEY PLAY?
The Foundation does not control the ibbCoin or ibbCoin DAO. The Foundation consists of a board of directors that exists solely to oversee the decisions of the ibbCoin DAO, as well as a third-party project management team responsible for ensuring that the decisions of the ibbCoin DAO are implemented.

A decentralized autonomous organization (DAO) is the best way to give every member of the community a vote on important decisions, whether it's a decision to fund a technical upgrade or a new idea. But the reality is that today a DAO cannot sign a lease or hire people or manufacture products or whatever the community decides to do on its own. The Foundation is responsible for the day-to-day administration, accounting, project management and other tasks that ensure the ibbCoin DAO community has the support they need to make their ideas come true.

IBBCOIN MANAGERS

FOUNDER OF GÜNEY (İBBVAKFI)


GUL (IBBVAKFI) CHAIRMAN OF THE BOARD


KETY (IBBVAKFI) BOARD MEMBER

Copyright © 2022 İBBCOIN - All Rights Reserved.